Move fixtures to data file tests
